Move CreatedModifiedFactory
