+++ v1/blocks/factories/block.py
@@ -1,7 +1,7 @@
 from factory import Faker
 from thenewboston.constants.network import BALANCE_LOCK_LENGTH, SIGNATURE_LENGTH, VERIFY_KEY_LENGTH
 
-from .created_modified import CreatedModifiedFactory
+from v1.third_party.thenewboston.factories.created_modified import CreatedModifiedFactory
 from ..models.block import Block
