+++ v1/invalid_blocks/factories/invalid_block.py
@@ -2,7 +2,7 @@ from factory import Faker, SubFactory
 from thenewboston.constants.network import BLOCK_IDENTIFIER_LENGTH
 
 from v1.blocks.factories.block import BlockFactory
-from v1.blocks.factories.created_modified import CreatedModifiedFactory
+from v1.third_party.thenewboston.factories.created_modified import CreatedModifiedFactory
 from v1.validators.factories.validator import ValidatorFactory
 from ..models.invalid_block import InvalidBlock
